_API.md fix for grewal
it seems internet dies daily...
i am trying to push fix for the API.md (renamed to _API.md) generation.
It used to generate the API.md then the help.json files, so you would
have to run it twice for the API.md to get the updates. with the fix,
it should work in one pass as I generate the help. json and then the
API.md uses the most recent help.json