txdetails and helper functions, proposed fix for #59
https://github.com/jl777/SuperNET/issues/59 required retrieving
additional details from transaction info RPC. both getrawtransaction
and get transaction use the same code path and list transactions also
uses the get transaction path, so all paths should now have the
txdetails fields:

     "category": "receive",
     "amount": 0.50000000,
     "confirmations": 24466,
     "blockhash":
"00000000000000000517ce625737579f91162c46ad9eaccad0f52ca13715b156",
     "blocktime": 1448045745,

     "blockindex": 78,
     "label": "",

the block index and label fields are currently not supported. Also
since iguana has data for all addresses, not just wallet addresses and
including ‘push’ addresses, the ‘category’ field has additional return
possibilities:

‘isp2sh’ is returned for any ‘p2sh’ address, regardless of whether some
or all of the required privkeys to spend the ‘p2sh’ address are in the
wallet. to determine partial or full signature capability would require
interpreting the script in place, and the returned information would
need to indicate conditional states, i.e. cltv, multisig, if/else, etc.
so this part is a won’t fix

for non-p2sh addresses, if it is in the wallet it will return “receive”
or “send” depending on whether the txid/vout outpoint has been spent or
not. for addresses no in the wallet, it will return “unspent” or
“spent”.

Note that a locked wallet might or might not get as far as calling this
txdetails, but if it did, it would return “unspent”/“spent” even for an
address in the wallet as it has no way to know if it is indeed a wallet
address