sendtoaddress fixed
reworked internals for basilisk/history api by using iguana/balance api
and sendtoaddress uses this data. tested finding and using utxo in
completed bundles, and also in the RT bundle. Left to test are
detecting a spend of RT bundle in the RT bundle. also amounts and
txfees are a bit off, but basic sendtoaddress should work in all cases
other than bugs

left to do is to lock utxo after being used so multiple tx per block
can be sent, otherwise the same utxo could be selected and that is
invalid