setprice 0
added a destpubkey field to the buy/sell trade api, this allows to
ordermatch to a specific entry in the orderbook. I will also be adding
a "depth" field to the orderbook, which will just be the sum of the max
size utxo from each orderbook entry. alternatively I could sum the
total balance at each orderbook entry, though at any given time you can
only buy one utxo. i guess it should be easy enough for the GUI to sum
the max utxo size, so I will make the depth the theoretical max depth,
summing the entire balances

I also now added automated broadcast of any setprices, which will occur
automatically when you do a buy/sell for the coin you are buying with,
as long as you are not using electrum. to be a bob, you need the native
coin. with the pruning of the orderbook to most recent 2 minutes, it
required the setprice to be called regularly. I have internalized that
so a single setprice is all that is needed. If you want to "cancel" it
you can setprice to 0