Bindgen from build.rs
Running bindgen from build.rs allows us to build the Rust version of the code out of the box, with "cargo build" and nothing else. This is particularly important in order for the Rust Language Server in the IDE to work.
Plus it gives us extra flexibility, which already helps with MM_VERSION.