Local windows build
The project now builds on my Windows workstation. Had to opt out from outdated HunterGate and DownloadProject includes (in the future if we'd need to download a library I'd like to do that in Rust, more reliable that way).

Binary renamed to `mm2` because it affects the namespace we're getting from the debugging information in the stack trace.

`cargo build` will run CMake now, hopefully allowing developers to Rust-build the project out of the box (provided that the build dependencies, such as the Visual Studio Build Tools, are there).

CIs now run the Rust tests.