Move Rust sources into separate folder
Rust sources were initially placed according to the existing C code packaging layout and near to the ported C sources (aka the "package by feature" approach) in order to better match the pre-existing knowledge, draw connections, find related files, etc.

But we also want the C code to automatically rebuild whenever it changes, and this is somewhat problematic to do with `rerun-if-changed` (we could make build.rs share a full list of C files with Cargo, but this list will also overflow the limited CI logs). It might be easier to simply move the Rust files into a separate folder, allowing us to trigger C rebuilds with a list of C directories and not just individual C files.