Bind to the specified IP address
Binding to "0.0.0.0" is a good default, buf if the IP address was specified on the command-line then we should only bind there, as requested. In particular, this allows us to run multiple MM instances on the same host.

The command-line "myipaddr" option is special in that it was introduced in MM2 and acting on it imposes no risk of breaking compatibility with MM1. I think we should only start making the breaking changes when the GUI team takes MM2 into their loop.