Optionally test MM in a local thread
This allows developers to more easily debug (with `gdb` or IDE debugger) the MarketMaker in the context created by an integration test.

In the future we hope to be able to run and test multiple MarketMaker instances in the same process.

The new command-line option `log` helps there, it will allow different MarketMaker instances to work with different log files.

Also in this commit the alphanumeric sorting is restored in the command-line options. Plus we might want to fit 123 columns (123 should be easy to remember).