Simplify crash handler tests
Got a fork in the ass. As we add new tests and with them executed in parallel, testing the access violation handler in a forked process began to cause problems, particularly on MacOS.
This commit triggers the access violation in a spawned process, which on some operating systems should be more reliable.