Port "stop"; refactor the RPC dispatcher
The "stop" method is fully ported to Rust.

Dispatcher is made into an explicit separate method.

Dispatcher returns a future and not a Hyper response, allowing the handlers to produce any HTTP response (changing HTTP status, headers, etc).

Error handling is simplified, making dispatcher more readable and maintainable. It should be okay, IMO, if dispatcher errors generate the same HTTP status (500). HTTP 400 means that there is no such page, using it for missing handlers *withing* an existing page doesn't match the HTTP spec very well, IIRC.

The lock, emulating the single-threaded execution for the old C stats_JSON code, was missing from the implementation. The old C code isn't thread-safe and we've discussed before that in a multi-threading dispatcher it must be behind a lock. I've added the misssing lock.