Port LP_autoprice
I've tried to improve readability of the resulting `lp_autoprice` in order to make the rest of the pricing code easier to grasp (when working with various parameters in the code it's important to know where they're coming from).

The `HyRes` RPC response has been moved to `helpers` in order not to introduce unnecessary layers and complexity and to allow subcrate handlers the full flexibility in generating the response `Future`.