Fix Windows stdout overlap
Printing of the debugging and tracing information from Rust is now performed via the `log!` macro, which for now uses the C `printf` in order to access the standard output, which works as a workaround to C and Rust standard outputs (and standard error streams) overlapping on Windows.

I expect there will be a time when a strong majority of the debugging and trancing information will be coming from Rust and not C. We can then flip this and make the C code print the debugging information through Rust.