Move CORS away from the context constructor
We might need to generate the CORS headers dynamically in the future, depending on the incoming Origin or encryption for example, so it might be better not to pregenerate the headers.

Moving the CORS away from the MmCtx constructor simplifies the MmCtx creation in unit tests, etc.

This commit is also a WIP on the peer communication API.