Support overwriting an existing seed-salt with a smaller value
When overwriting seed-salt values (wether intentionally or due to chance), we might encounter a situation when the first fetch of the first chunk returned a larger number_of_chunks than the consequent fetches. Corrent action then is to reduce the number of chunks fetched, in order not to fetch and merge the churned/outdated chunks.