Upgrade the peers API to Rust idioms
Now that Bob and Alice loops are ported we can use the `peers` crate direcly.

The socket and clock abstractions were removed in favor of the more flexible and domain-specific `subject` identification.

Instead of continuously pinging the `peers_recv_compat` the `Stream` and `Future` instances are now used to "hold" the storage and retrieval operations.