Move the tests into a separate file
I'm using the `pub mod coins_tests` form (suitable for linking the tests into the `mm2_tests` bundle) because that way the test development is properly enhanced in VSCode, with RLS providing error indication, type inference and code completion in the tests.
(Also makes them easier to find and to run, IMHO).
#233