Keep market operations in a separate trait
Discussing the latest patches we figured that the financial / market operations implemented by coins are different from the MM-specific knobs and operations and that we'd like to separate them in order to keep things clear.
Here this is done by keeping the financial and market operations in the MarketCoinOps subtrait.
The plan is for MmCoin to later grow other subtraits, related to MM-specific configuration and such.
#233