Port lp_coininit, WIP
Coin initialization has moved to Rust (lp_coininit). And MmCoinEnum is created in the context right away.

In this commit we're temporarily ignoring the coin configuration values.

Moving the iguana_info to Rust and pinning it has solved some of the memory corruption that I've been observing on the Jenkins server, but the trade test isn't working for me yet (might be because the coin configuration is not loaded by the new code yet).

Part of #233