Port LP_coincreate
Also fixes the bothersome memory corruption, the one that prevented me from testing the trade on the Jenkins server: there was a different version of `struct iguana_info` that was randomly used in place of the correct version. I was lucky to catch this other version, and with the same "corrupted" values that I've seen on the server, in the VSCode debugger on Windows. Solved by removing the duplicate from the headers.