Use direct communication in send and recv
UDP ping packets are used as in parallel with DHT, allowing the packages to be delivered much faster whenever there is a NAT/firewall hole for those packets to squeeze through.

Direct communication channel is established with `investigate_peer`, no automatic discovery YET.

The sending effort is now properly continuous, renewing the DHT chunks until the send handler is dropped.