Fallback to 0.0.0.0 and 127.0.0.1
If the user has given a particular IP, we'd like to assume that she knows what she's doing, expects that IP to be used and be promptly informed with a failure if that IP doesn't work.

If, on the other hand, we're detecting the IP all by ourselves, then we should account for NAT (when the detected outer IP doesn't exist on a connected interface) or the possibility of another MM2 instance already running on the host.

#273