Use the SWAP UUID in the subject; debug logs
This commit prints DHT stores and DHT loads into the log, identified by chunk salt and number. We want them in order to debug the possibility of the DHT failures. We'll probably turn them back off when the DHT communication is stable.

Also the chunk salt is now a checksum, making it safe from long and/or binary subjects. Large salt might have been the cause of some of the DHT failures we've seen during the testing: the size of the salt is limited (BEP 44 mentions error code 207), in underspecified fashion.

#285