Share the original ask.price as is
The 1 / ask.price division can not be precise because the rational numbers can not be represented as decimals without the infinite decimal support.
We'd like to preserve the original price when it is available as it might help with precisely matching the price in the GUIs and with troubleshooting price issues.

#495