Use Constructible
To port the remaining global (singleton) C structures into the MmCtx instance fields we need a convenient abstraction.
`Mutex`, `Option` and `Atomic` are falling short of it because they wouldn't capture the idea of delayed initialization. I mean, we have fields which are, once initialized, essentially immutable, and having this concept of delayed initialization captured in the corresponding abtraction really clears things up (in comparison to using mutables).
I've been missing such an abstraction for quite some time, actually.