Use a dedicated thread in Timer
According to the stack trace at https://github.com/ca333/komodoDEX/issues/360#issuecomment-531699919 we might hit the thread starvation sometimes. Using but a single thread in the Timer should help to avoid that.

I've implemented the `spawn_after` interface that we have in the WASM executor as it might help with the porting effort.

Known alternatives are:
1) To reuse the newer tokio timers. It isn't clear how to make them work though, for me they'd always fail, even on the tokio runtime, as we need to pin the tokio timer service to the runtime and it isn't currently docummented. And they're using more then one thread IIRC.
2) Use the outdated timer crates (tokio timers 1, futures-timer). They are targeted at the old pre-async/await version of the futures.
3) Use OS timers. Such timers aren't portable, harder to implement and more risky.